little Housekeeping

+++ roboshop/common.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/bash
 
+USER_ID=$(id -u)
+if [ $USER_ID -ne 0 ]; then
+  echo -e "\e[32mYou should be a root user or sudo user to run this command\e[0m"
+  exit 2
+fi
+
 LOG=/tmp/roboshop.log
 rm -f $LOG
 
@@ -15,9 +21,3 @@ STAT_CHECK(){
 PRINT(){
    echo -n -e "$1\t\t"
 }
-
-USER_ID=$(id -u)
-if [ $USER_ID -ne 0 ]; then
-  echo -e "\e[32mYou should be a root user or sudo user to run this command\e[0m"
-  exit 2
-fi
